# api-spec.yaml
openapi: 3.0.3
info:
  title: Next Loyalty Program API
  description: 'Unified Loyalty Program API for Next plc omnichannel retail.

    Supports customer registration, points earning/redemption, tier management,

    notifications, staff operations, and admin configuration.

    '
  version: 1.0.0
  contact:
    name: Next Loyalty Engineering
    email: loyalty-engineering@next.co.uk
servers:
- url: https://loyalty.next.co.uk/api/v1
  description: Production
- url: https://staging.loyalty.next.co.uk/api/v1
  description: Staging
- url: http://localhost:3000/api/v1
  description: Local Development
security:
- bearerAuth: []
...
paths:
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      summary: Get admin dashboard metrics
      description: Overview metrics including active members, sign-ups, redemption
        rate, and points liability. Requires loyalty_admin role.
      tags:
      - Admin
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  activeMembers:
                    type: integer
                    example: 85000
                  signUpsThisMonth:
                    type: integer
                    example: 4200
                  redemptionRate:
                    type: number
                    format: double
                    example: 32.5
                  totalPointsLiability:
                    type: integer
                    example: 12500000
                  activeCampaigns:
                    type: integer
                    example: 3
                  transactionsToday:
                    type: integer
                    example: 1850
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/config/rates:
    get:
      operationId: getConfigRates
      summary: Get loyalty program rates
      description: Retrieve current accrual rate, redemption rate, and rule configuration.
      tags:
      - Admin
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Current rates
          content:
            application/json:
              schema:
                type: object
                properties:
                  accrualRate:
                    type: number
                    format: double
                    description: Points per £1 spent
                    example: 1.0
                  redemptionRate:
                    type: number
                    format: double
                    description: Points per £1 discount
                    example: 20.0
                  minimumRedemption:
                    type: integer
                    description: Minimum points to redeem
                    example: 100
                  maxDiscountPercentage:
                    type: number
                    format: double
                    description: Maximum discount as percentage of order total
                    example: 50.0
                  expirationMonths:
                    type: integer
                    description: Months of inactivity before expiration (0 = never)
                    example: 12
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: updateConfigRates
      summary: Update loyalty program rates
      description: Update accrual/redemption rates and rules. Changes take effect
        immediately. Audit-logged.
      tags:
      - Admin
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accrualRate:
                  type: number
                  format: double
                redemptionRate:
                  type: number
                  format: double
                minimumRedemption:
                  type: integer
                maxDiscountPercentage:
                  type: number
                  format: double
                expirationMonths:
                  type: integer
      responses:
        '200':
          description: Rates updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/campaigns:
    get:
      operationId: listCampaigns
      summary: List promotional campaigns
      description: Retrieve all campaigns with status and performance metrics.
      tags:
      - Admin
      security:
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - draft
          - active
          - paused
          - completed
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createCampaign
      summary: Create promotional campaign
      description: Create a new bonus points campaign with eligibility rules and budget.
      tags:
      - Admin
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - type
              - value
              - startDate
              - endDate
              - eligibility
              properties:
                name:
                  type: string
                  example: Summer Double Points
                type:
                  type: string
                  enum:
                  - multiplier
                  - fixed_bonus
                  - category_bonus
                value:
                  type: number
                  format: double
                  description: Multiplier (e.g. 2.0) or fixed bonus points
                  example: 2.0
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
                eligibility:
                  type: string
                  enum:
                  - all_customers
                  - tier_minimum
                  - new_members_only
                minimumTier:
                  type: string
                  nullable: true
                  description: Required if eligibility is tier_minimum
                applicableCategories:
                  type: array
                  items:
                    type: string
                maxBudget:
                  type: integer
                  nullable: true
                  description: Maximum total bonus points to award (null = unlimited)
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/campaigns/{id}:
    put:
      operationId: updateCampaign
      summary: Update promotional campaign
      description: Update an existing campaign. Active campaigns can be paused or
        have end date extended.
      tags:
      - Admin
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  type: string
                  enum:
                  - draft
                  - active
                  - paused
                  - completed
                endDate:
                  type: string
                  format: date-time
                maxBudget:
                  type: integer
                  nullable: true
      responses:
        '200':
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/audit:
    get:
      operationId: getAuditLog
      summary: Query audit trail
      description: Paginated, filterable audit log of all state-changing operations.
        Immutable, retained 7 years.
      tags:
      - Admin
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/PageNumber'
      - $ref: '#/components/parameters/PageSize'
      - name: customerId
        in: query
        schema:
          type: string
          format: uuid
      - name: action
        in: query
        schema:
          type: string
      - name: actorType
        in: query
        schema:
          type: string
          enum:
          - customer
          - staff
          - system
      - name: fromDate
        in: query
        schema:
          type: string
          format: date-time
      - name: toDate
        in: query
        schema:
          type: string
          format: date-time
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEntry'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (15-minute expiry)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for POS terminals and e-commerce integrations
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Unique key to ensure idempotent processing of write operations
    PageNumber:
      name: pageNumber
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
...
    Campaign:
      type: object
      properties:
        campaignId:
          type: string
          format: uuid
        name:
          type: string
          example: Summer Double Points
        type:
          type: string
          enum:
          - multiplier
          - fixed_bonus
          - category_bonus
        value:
          type: number
          format: double
          example: 2.0
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        eligibility:
          type: string
          enum:
          - all_customers
          - tier_minimum
          - new_members_only
        minimumTier:
          type: string
          nullable: true
        applicableCategories:
          type: array
          items:
            type: string
          nullable: true
        maxBudget:
          type: integer
          nullable: true
        budgetUsed:
          type: integer
          example: 45000
        status:
          type: string
          enum:
          - draft
          - active
          - paused
          - completed
        createdAt:
          type: string
          format: date-time
      example:
        campaignId: e5f6a7b8-c9d0-1234-ef56-789012345678
        name: Summer Double Points
        type: multiplier
        value: 2.0
        startDate: '2026-06-01T00:00:00Z'
        endDate: '2026-06-30T23:59:59Z'
        eligibility: all_customers
        minimumTier: null
        applicableCategories: null
        maxBudget: 500000
        budgetUsed: 45000
        status: active
        createdAt: '2026-05-15T09:00:00Z'
    AuditEntry:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        actorType:
          type: string
          enum:
          - customer
          - staff
          - system
        actorId:
          type: string
          format: uuid
        action:
          type: string
          example: points_earned
        entityType:
          type: string
          example: points_ledger
        entityId:
          type: string
          format: uuid
        beforeState:
          type: object
          nullable: true
        afterState:
          type: object
          nullable: true
        sourceSystem:
          type: string
          enum:
          - app
          - web
          - pos
          - admin
          - system
        ipAddress:
          type: string
          example: 192.168.1.100
        correlationId:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 156
        totalPages:
          type: integer
          example: 8
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: https://loyalty.next.co.uk/problems/validation-error
        title:
          type: string
          description: Short human-readable summary
          example: Validation Error
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
          example: The 'email' field must be a valid email address
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence
          example: /api/v1/auth/register
        traceId:
          type: string
          description: Correlation ID for distributed tracing
          example: abc123def456
        errors:
          type: array
          description: Detailed field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Must be a valid email address
              code:
                type: string
                example: INVALID_FORMAT
      example:
        type: https://loyalty.next.co.uk/problems/validation-error
        title: Validation Error
        status: 422
        detail: One or more fields failed validation
        instance: /api/v1/auth/register
        traceId: abc123def456
        errors:
        - field: email
          message: Must be a valid email address
          code: INVALID_FORMAT
...
  responses:
    BadRequest:
      description: Bad Request — malformed request syntax or invalid parameters
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/bad-request
            title: Bad Request
            status: 400
            detail: Request body is malformed or missing required fields
            traceId: trace-001
    Unauthorized:
      description: Unauthorized — missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/unauthorized
            title: Unauthorized
            status: 401
            detail: Access token is missing, expired, or invalid
            traceId: trace-002
    Forbidden:
      description: Forbidden — authenticated but insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            traceId: trace-003
    NotFound:
      description: Not Found — requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found
            traceId: trace-004
    UnprocessableEntity:
      description: Unprocessable Entity — request understood but validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/validation-error
            title: Validation Error
            status: 422
            detail: One or more fields failed validation
            traceId: trace-005
            errors:
            - field: email
              message: Email already registered
              code: DUPLICATE
    InternalServerError:
      description: Internal Server Error — unexpected server-side failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://loyalty.next.co.uk/problems/internal-error
            title: Internal Server Error
            status: 500
            detail: An unexpected error occurred. Please try again later.
            traceId: trace-006